:sparkles: Main Template Module
