+++ src/components/molecules/Appbar/styles.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled'
+import { Toolbar } from '@mui/material';
+
+export const StyledToolbar = styled(Toolbar)`
+    background-color: #fafafa;
+    display: flex;
+    justify-content: space-between;
+    height: 122px;
+`
