--- src/components/pages/products/productOverview.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ProductsOverviewPage: React.FC = () => {
-  return <span>Products Overview Page</span>;
-};
-
-export default ProductsOverviewPage;
