--- src/components/pages/products/productsCardPage.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ProductsCartPage: React.FC = () => {
-  return <span>Products Cart Page</span>;
-};
-
-export default ProductsCartPage;
