--- src/components/pages/products/productsPage.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ProductsPage: React.FC = () => {
-  return <span>Products Page</span>;
-};
-
-export default ProductsPage;
