+++ src/pages/products/productOverview.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ProductsOverviewPage: React.FC = () => {
+  return <span>Products Overview Page</span>;
+};
+
+export default ProductsOverviewPage;
