+++ src/pages/products/products.mock.ts
@@ -1,0 +1,402 @@
+export const products = [
+  {
+    id: "1",
+    createdAt: "2019-09-02T12:58:54.103Z",
+    name: "Rustic Metal Fish",
+    price: "289.00",
+    image: "http://lorempixel.com/640/480/food",
+    stock: 65171,
+  },
+  {
+    id: "2",
+    createdAt: "2019-09-02T07:59:58.181Z",
+    name: "Sleek Wooden Soap",
+    price: "430.00",
+    image: "http://lorempixel.com/640/480/transport",
+    stock: 91260,
+  },
+  {
+    id: "3",
+    createdAt: "2019-09-02T22:14:05.454Z",
+    name: "Small Cotton Shoes",
+    price: "993.00",
+    image: "http://lorempixel.com/640/480/sports",
+    stock: 36608,
+  },
+  {
+    id: "4",
+    createdAt: "2019-09-02T07:36:56.139Z",
+    name: "Ergonomic Frozen Towels",
+    price: "259.00",
+    image: "http://lorempixel.com/640/480/nightlife",
+    stock: 92065,
+  },
+  {
+    id: "5",
+    createdAt: "2019-09-02T05:18:30.393Z",
+    name: "Awesome Metal Pants",
+    price: "69.00",
+    image: "http://lorempixel.com/640/480/sports",
+    stock: 54930,
+  },
+  {
+    id: "6",
+    createdAt: "2019-09-02T16:22:03.725Z",
+    name: "Intelligent Cotton Bacon",
+    price: "14.00",
+    image: "http://lorempixel.com/640/480/fashion",
+    stock: 88996,
+  },
+  {
+    id: "7",
+    createdAt: "2019-09-02T02:26:28.018Z",
+    name: "Ergonomic Plastic Tuna",
+    price: "477.00",
+    image: "http://lorempixel.com/640/480/food",
+    stock: 73117,
+  },
+  {
+    id: "8",
+    createdAt: "2019-09-02T14:25:31.508Z",
+    name: "Intelligent Soft Keyboard",
+    price: "764.00",
+    image: "http://lorempixel.com/640/480/nature",
+    stock: 14875,
+  },
+  {
+    id: "9",
+    createdAt: "2019-09-02T11:52:19.177Z",
+    name: "Awesome Cotton Keyboard",
+    price: "575.00",
+    image: "http://lorempixel.com/640/480/transport",
+    stock: 55458,
+  },
+  {
+    id: "10",
+    createdAt: "2019-09-02T02:17:08.337Z",
+    name: "Small Fresh Shoes",
+    price: "67.00",
+    image: "http://lorempixel.com/640/480/fashion",
+    stock: 38268,
+  },
+  {
+    id: "11",
+    createdAt: "2019-09-02T04:41:05.149Z",
+    name: "Fantastic Metal Cheese",
+    price: "312.00",
+    image: "http://lorempixel.com/640/480/technics",
+    stock: 43921,
+  },
+  {
+    id: "12",
+    createdAt: "2019-09-02T12:39:09.964Z",
+    name: "Generic Granite Bike",
+    price: "345.00",
+    image: "http://lorempixel.com/640/480/people",
+    stock: 77846,
+  },
+  {
+    id: "13",
+    createdAt: "2019-09-02T06:35:40.678Z",
+    name: "Tasty Granite Computer",
+    price: "585.00",
+    image: "http://lorempixel.com/640/480/business",
+    stock: 67599,
+  },
+  {
+    id: "14",
+    createdAt: "2019-09-02T00:51:18.473Z",
+    name: "Rustic Plastic Keyboard",
+    price: "567.00",
+    image: "http://lorempixel.com/640/480/nightlife",
+    stock: 15619,
+  },
+  {
+    id: "15",
+    createdAt: "2019-09-02T13:03:07.618Z",
+    name: "Sleek Steel Hat",
+    price: "73.00",
+    image: "http://lorempixel.com/640/480/nature",
+    stock: 3784,
+  },
+  {
+    id: "16",
+    createdAt: "2019-09-02T03:28:32.706Z",
+    name: "Gorgeous Metal Tuna",
+    price: "966.00",
+    image: "http://lorempixel.com/640/480/city",
+    stock: 39055,
+  },
+  {
+    id: "17",
+    createdAt: "2019-09-02T04:04:19.662Z",
+    name: "Handmade Rubber Pants",
+    price: "334.00",
+    image: "http://lorempixel.com/640/480/city",
+    stock: 12305,
+  },
+  {
+    id: "18",
+    createdAt: "2019-09-02T21:09:28.496Z",
+    name: "Gorgeous Cotton Chips",
+    price: "5.00",
+    image: "http://lorempixel.com/640/480/nightlife",
+    stock: 40888,
+  },
+  {
+    id: "19",
+    createdAt: "2019-09-02T20:26:48.676Z",
+    name: "Unbranded Frozen Shoes",
+    price: "956.00",
+    image: "http://lorempixel.com/640/480/nature",
+    stock: 83757,
+  },
+  {
+    id: "20",
+    createdAt: "2019-09-02T13:55:49.480Z",
+    name: "Awesome Granite Ball",
+    price: "523.00",
+    image: "http://lorempixel.com/640/480/transport",
+    stock: 59087,
+  },
+  {
+    id: "21",
+    createdAt: "2019-09-02T07:46:29.266Z",
+    name: "Practical Plastic Shoes",
+    price: "103.00",
+    image: "http://lorempixel.com/640/480/sports",
+    stock: 59388,
+  },
+  {
+    id: "22",
+    createdAt: "2019-09-02T20:02:56.461Z",
+    name: "Sleek Soft Sausages",
+    price: "878.00",
+    image: "http://lorempixel.com/640/480/abstract",
+    stock: 61433,
+  },
+  {
+    id: "23",
+    createdAt: "2019-09-02T22:58:53.306Z",
+    name: "Small Cotton Pizza",
+    price: "863.00",
+    image: "http://lorempixel.com/640/480/fashion",
+    stock: 96285,
+  },
+  {
+    id: "24",
+    createdAt: "2019-09-02T08:02:02.139Z",
+    name: "Small Granite Towels",
+    price: "909.00",
+    image: "http://lorempixel.com/640/480/nightlife",
+    stock: 66317,
+  },
+  {
+    id: "25",
+    createdAt: "2019-09-02T07:48:54.872Z",
+    name: "Small Concrete Chair",
+    price: "978.00",
+    image: "http://lorempixel.com/640/480/fashion",
+    stock: 93019,
+  },
+  {
+    id: "26",
+    createdAt: "2019-09-02T19:20:49.940Z",
+    name: "Small Rubber Bike",
+    price: "734.00",
+    image: "http://lorempixel.com/640/480/cats",
+    stock: 32508,
+  },
+  {
+    id: "27",
+    createdAt: "2019-09-02T12:54:59.331Z",
+    name: "Awesome Plastic Mouse",
+    price: "375.00",
+    image: "http://lorempixel.com/640/480/technics",
+    stock: 83785,
+  },
+  {
+    id: "28",
+    createdAt: "2019-09-02T01:24:49.343Z",
+    name: "Unbranded Granite Ball",
+    price: "799.00",
+    image: "http://lorempixel.com/640/480/fashion",
+    stock: 10590,
+  },
+  {
+    id: "29",
+    createdAt: "2019-09-02T15:49:48.296Z",
+    name: "Handcrafted Concrete Car",
+    price: "323.00",
+    image: "http://lorempixel.com/640/480/food",
+    stock: 33103,
+  },
+  {
+    id: "30",
+    createdAt: "2019-09-02T08:08:23.685Z",
+    name: "Tasty Wooden Keyboard",
+    price: "662.00",
+    image: "http://lorempixel.com/640/480/food",
+    stock: 84332,
+  },
+  {
+    id: "31",
+    createdAt: "2019-09-02T18:39:51.480Z",
+    name: "Small Cotton Hat",
+    price: "933.00",
+    image: "http://lorempixel.com/640/480/people",
+    stock: 8714,
+  },
+  {
+    id: "32",
+    createdAt: "2019-09-02T04:13:14.888Z",
+    name: "Licensed Cotton Towels",
+    price: "48.00",
+    image: "http://lorempixel.com/640/480/people",
+    stock: 9829,
+  },
+  {
+    id: "33",
+    createdAt: "2019-09-02T03:34:04.876Z",
+    name: "Rustic Rubber Keyboard",
+    price: "18.00",
+    image: "http://lorempixel.com/640/480/nature",
+    stock: 21040,
+  },
+  {
+    id: "34",
+    createdAt: "2019-09-02T12:37:28.639Z",
+    name: "Gorgeous Plastic Chicken",
+    price: "261.00",
+    image: "http://lorempixel.com/640/480/fashion",
+    stock: 44945,
+  },
+  {
+    id: "35",
+    createdAt: "2019-09-02T17:55:12.895Z",
+    name: "Small Rubber Chair",
+    price: "539.00",
+    image: "http://lorempixel.com/640/480/sports",
+    stock: 40411,
+  },
+  {
+    id: "36",
+    createdAt: "2019-09-02T20:17:40.306Z",
+    name: "Intelligent Granite Pizza",
+    price: "221.00",
+    image: "http://lorempixel.com/640/480/nature",
+    stock: 29440,
+  },
+  {
+    id: "37",
+    createdAt: "2019-09-02T12:36:49.977Z",
+    name: "Small Granite Towels",
+    price: "639.00",
+    image: "http://lorempixel.com/640/480/nature",
+    stock: 6285,
+  },
+  {
+    id: "38",
+    createdAt: "2019-09-02T03:28:40.423Z",
+    name: "Sleek Metal Chips",
+    price: "115.00",
+    image: "http://lorempixel.com/640/480/cats",
+    stock: 19076,
+  },
+  {
+    id: "39",
+    createdAt: "2019-09-02T11:02:15.185Z",
+    name: "Handcrafted Steel Chicken",
+    price: "385.00",
+    image: "http://lorempixel.com/640/480/business",
+    stock: 72860,
+  },
+  {
+    id: "40",
+    createdAt: "2019-09-02T05:17:01.690Z",
+    name: "Incredible Wooden Mouse",
+    price: "527.00",
+    image: "http://lorempixel.com/640/480/transport",
+    stock: 8817,
+  },
+  {
+    id: "41",
+    createdAt: "2019-09-02T22:02:50.052Z",
+    name: "Licensed Wooden Chicken",
+    price: "552.00",
+    image: "http://lorempixel.com/640/480/cats",
+    stock: 99300,
+  },
+  {
+    id: "42",
+    createdAt: "2019-09-02T13:06:02.107Z",
+    name: "Ergonomic Wooden Gloves",
+    price: "84.00",
+    image: "http://lorempixel.com/640/480/business",
+    stock: 64007,
+  },
+  {
+    id: "43",
+    createdAt: "2019-09-02T01:52:34.371Z",
+    name: "Generic Rubber Shoes",
+    price: "233.00",
+    image: "http://lorempixel.com/640/480/fashion",
+    stock: 56922,
+  },
+  {
+    id: "44",
+    createdAt: "2019-09-02T04:35:26.088Z",
+    name: "Practical Fresh Chair",
+    price: "706.00",
+    image: "http://lorempixel.com/640/480/cats",
+    stock: 33907,
+  },
+  {
+    id: "45",
+    createdAt: "2019-09-02T04:53:47.174Z",
+    name: "Handcrafted Frozen Keyboard",
+    price: "698.00",
+    image: "http://lorempixel.com/640/480/nightlife",
+    stock: 85011,
+  },
+  {
+    id: "46",
+    createdAt: "2019-09-02T15:56:19.734Z",
+    name: "Incredible Steel Chair",
+    price: "175.00",
+    image: "http://lorempixel.com/640/480/nightlife",
+    stock: 38658,
+  },
+  {
+    id: "47",
+    createdAt: "2019-09-02T03:48:29.016Z",
+    name: "Tasty Soft Chicken",
+    price: "514.00",
+    image: "http://lorempixel.com/640/480/technics",
+    stock: 22612,
+  },
+  {
+    id: "48",
+    createdAt: "2019-09-02T13:24:55.100Z",
+    name: "Intelligent Plastic Chair",
+    price: "15.00",
+    image: "http://lorempixel.com/640/480/people",
+    stock: 99397,
+  },
+  {
+    id: "49",
+    createdAt: "2019-09-02T06:45:33.429Z",
+    name: "Fantastic Wooden Gloves",
+    price: "387.00",
+    image: "http://lorempixel.com/640/480/abstract",
+    stock: 32266,
+  },
+  {
+    id: "50",
+    createdAt: "2019-09-02T14:00:40.401Z",
+    name: "Awesome Plastic Sausages",
+    price: "495.00",
+    image: "http://lorempixel.com/640/480/nightlife",
+    stock: 27837,
+  },
+];
