+++ src/pages/products/productsCardPage.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ProductsCartPage: React.FC = () => {
+  return <span>Products Cart Page</span>;
+};
+
+export default ProductsCartPage;
