+++ src/pages/products/productsPage.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import MainTemplate from "../../components/templates/mainTemplate";
+
+const ProductsPage: React.FC = () => {
+  return <MainTemplate>Olha que tal</MainTemplate>;
+};
+
+export default ProductsPage;
